pyproject.toml
[build-system]
build-backend = "uv_build"
requires = ["uv_build"]

[dependency-groups]
dev = ["pyright", "pytest", "pytest-asyncio", "ruff"]

[project]
name = "pydantic-settings-op"
version = "0.1.0"
description = "1Password secrets source for pydantic-settings"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Jacob Hayes" }]
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Topic :: Security",
  "Typing :: Typed",
]
dependencies = [
    "onepassword-sdk>=0.4.1b1",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.urls]
Homepage = "https://github.com/JacobHayes/pydantic-settings-op"
Repository = "https://github.com/JacobHayes/pydantic-settings-op"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.12"
reportMissingTypeStubs = false
typeCheckingMode = "strict"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]
